Give the OptionSetFlags word one home in `types::option`
`OPT_LOCAL` was re-declared 31 times, `OPT_GLOBAL` 13, and the other five
members of the same flag word 2-4 times each: 59 declarations of seven
constants across 33 files. Every value agreed with upstream's
`OptionSetFlags` enum in `option.h`, so the dedupe is value-preserving.

Deduping one member of a flag word in isolation would have been churn the
flag-word slice immediately redid, so the whole word moved at once --
minus `OPT_NO_REDRAW` (0x40), which upstream defines and this tree has
never had a copy of because nothing uses it. The canonical type is
`OptionSetFlags = c_uint`, which is what the C enum is and what 33 of the
59 copies already said; the ~65 sites that had it as `c_int` take
`as c_int`. That is a holding position: the flag-word slice replaces the
alias with a `flag_set!` newtype and the casts go with it.

The `OPT_*` names that are *not* this word -- `OPT_HISTORY`,
`OPT_HIT_ENTER`, `OPT_PROGRESS`, `OPT_WAIT`, `OPT_SCOPES` -- share only
the prefix and were left alone.

`--allow-growth`: +7 unchecked lines, all rustfmt reflow around the new
casts.

+++ crates/nvim/src/eval/vars/scoped.rs
@@ -12,7 +12,7 @@ use core::ffi::{c_char, c_int};
 use core::ptr;
 
 use super::*;
-use crate::types::{NUL, OK};
+use crate::types::{NUL, OK, OPT_LOCAL};
 
 /// The zeroed `switchwin_T` [`switch_win`] fills in.
 const SWITCHWIN_INITIAL_VALUE: switchwin_T = switchwin_T {
@@ -300,7 +300,7 @@ unsafe fn set_option_from_tv(varname: *const c_char, varp: *mut typval_T) {
         let mut error = false;
         let value = tv_to_optval(varp, opt_idx, varname, &raw mut error);
         if !error {
-            let errmsg = set_option_value_handle_tty(varname, opt_idx, value, OPT_LOCAL);
+            let errmsg = set_option_value_handle_tty(varname, opt_idx, value, OPT_LOCAL as c_int);
             if !errmsg.is_null() {
                 emsg(errmsg);
             }
